requirements.txt
langchain
transformers
faiss-cpu
pymilvus
pinecone-client
black
torch
tensorflow
torchvision
torchaudio
tf-keras
kaggle
pandas
scikit-learn
sentence_transformers
numpy
faiss-gpu
faiss-cpu
psutil
flask
langchain
openai
langchain-community
python-dotenv
tiktoken
langchain-openai